backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
networkx
pandas
google-adk
google-genai
sse-starlette
